Latest Website/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.103.2
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
httptools==0.6.0
idna==3.4
langchain==0.0.268
langsmith==0.0.42
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.0
openai==0.27.8
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
pypdf==3.15.2
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
PyYAML==6.0
regex==2022.7.9
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
striprtf==0.0.26
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
